break out dfs code into monolithic function and add support for dfs in get_links_and_load
